Replace App Content With Skeleton
Replace the default content of the application with a rough outline
of what's to come. Create react app provides starter content for the
site, but we don't want it. We want to put our own content in the site.

If we end up having a header, we'll probably want the header component
to be consistent across all "pages." I've set aside a spot for the
header
in the overall app. We can put the logic for switching between "pages"
below that.

I put "pages" in quotes since the "pages" will be components, and it's
technically a single page app that will switch the content of the page
depending on the situation.